+++ client/package.json
@@ -17,6 +17,7 @@
     "react": "^18.2.0",
     "react-archer": "^4.3.0",
     "react-dom": "^18.2.0",
+    "react-xarrows": "^2.0.2",
     "styled-components": "^6.1.8",
     "zustand": "^4.5.2"
   },
