+++ client/src/types/INodeItem.ts
@@ -14,5 +14,5 @@ export interface INodeItem {
   text: string
   position: IPosition
   shape?: TShape
-  targetId?: string
+  targetIds?: Array<string>
 }
